pyproject.toml
[tool.poetry]
name = "hh-ru-sql"
version = "0.1.0"
description = ""
authors = ["Alina Davydenko <Alinadavydenko620@gmail.com>"]
readme = "README.md"
packages = [{include = "hh"}]

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
